Add files via upload
The aim of this project was to find all CAS numbers, boiling points and melting using the main website described below.
As the site contains several "subsites" that direct you to the compound, another request was required. 
Some information might have been repeated because there was more than one "subsites" that direct you to the same information. 
Neither melting point nor boiling point were possible to find for some compounds because the site didn't have that information.
The huge difficulty was to access the correct information once those "subsites" (c_links) were not displayed in the same way.